new files on OOP added

+++ PrimeNumbers1.py
@@ -6,7 +6,7 @@ for i in range(3, 3000001):
     for j in primeList:
         if((i % j == 0)):
             break
-        elif(j >= int(sqrt(i))):
+        elif(j >= sqrt(i)):
             prime += 1
             primeList.append(i)
             break
